Restructure jobs and scripts
The previous structure did not clearly separate between configuration
for a single backup and configuration which may in theory be used by
multiple backup jobs.

This new structure clarifies this. Each configuration directory is for a
group of systemd units with the same instance name. This means that the
same configuration may be used by a filesystem and a postgresql database
backup. However, two separate filesystem backups will need two separate
instances.

As of now the newly structured postgresql backup remains untested. I
first need to find time to create the new testing infrastructure.